requirements.txt
fastapi==0.92.0
pydantic==1.10.4
Faker==16.6.1
# local development & testing
pytest==7.2.1
# hypercorn==0.14.3
uvicorn==0.20.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
pyathena==2.19.0
pylint==2.15.10
moto==4.1.1
python-dotenv==0.21.1
coverage==7.1.0
pyjwt==2.6.0
httpx==0.23.3
cachetools==5.2.0
pytest-cov==4.0.0
redis-py-cluster==2.1.3
cryptography==39.0.0
pynamodb==5.4.1
email-validator==1.3.1
black==22.12.0
pre-commit==3.0.2
moto[dynamodb]
jsonpickle==3.0.1
